+++ client/BlueMonkey/BlueMonkey.Application/App.xaml
@@ -2,6 +2,6 @@
 <prism:PrismApplication xmlns="http://xamarin.com/schemas/2014/forms"
                         xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
                         xmlns:prism="clr-namespace:Prism.Unity;assembly=Prism.Unity.Forms"
-                        x:Class="BlueMonkey.App">
+                        x:Class="BlueMonkey.Application.App">
 
 </prism:PrismApplication>
