+++ client/BlueMonkey/BlueMonkey.Application/Secrets_sample.cs
