+++ client/BlueMonkey/BlueMonkey.Application/app.config
