+++ client/BlueMonkey/BlueMonkey.Application/packages.config
@@ -1,0 +1,14 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="CommonServiceLocator" version="1.3" targetFramework="portable45-net45+win8" />
+  <package id="Microsoft.Azure.Mobile.Client" version="3.0.3" targetFramework="portable45-net45+win8" />
+  <package id="Microsoft.Bcl" version="1.1.10" targetFramework="portable45-net45+win8" />
+  <package id="Microsoft.Bcl.Build" version="1.0.21" targetFramework="portable45-net45+win8" />
+  <package id="Microsoft.Net.Http" version="2.2.29" targetFramework="portable45-net45+win8" />
+  <package id="Newtonsoft.Json" version="6.0.4" targetFramework="portable45-net45+win8" />
+  <package id="Prism.Core" version="6.3.0-pre1" targetFramework="portable45-net45+win8" />
+  <package id="Prism.Forms" version="6.3.0-pre1" targetFramework="portable45-net45+win8" />
+  <package id="Prism.Unity.Forms" version="6.3.0-pre1" targetFramework="portable45-net45+win8" />
+  <package id="Unity" version="4.0.1" targetFramework="portable45-net45+win8" />
+  <package id="Xamarin.Forms" version="2.3.2.127" targetFramework="portable45-net45+win8" />
+</packages>
