+++ client/BlueMonkey/BlueMonkey.Droid/MainActivity.cs
@@ -3,6 +3,7 @@
 using Android.App;
 using Android.Content.PM;
 using Android.OS;
+using BlueMonkey.Application;
 using BlueMonkey.ExpenseServices;
 using BlueMonkey.LoginService;
 #if Azure
