+++ client/BlueMonkey/BlueMonkey.Views/Behaviors/BehaviorBase.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Xamarin.Forms;
 
-namespace BlueMonkey
+namespace BlueMonkey.Views.Behaviors
 {
     public class BehaviorBase<T> : Behavior<T> where T : BindableObject
     {
