+++ client/BlueMonkey/BlueMonkey.Views/ExpenseListPage.xaml
@@ -2,7 +2,7 @@
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:prism="clr-namespace:Prism.Mvvm;assembly=Prism.Forms"
-             xmlns:blueMonkey="clr-namespace:BlueMonkey;assembly=BlueMonkey"
+             xmlns:behaviors="clr-namespace:BlueMonkey.Views.Behaviors;assembly=BlueMonkey.Views"
              prism:ViewModelLocator.AutowireViewModel="True"
              x:Class="BlueMonkey.Views.ExpenseListPage" Title="Expense List">
   <ContentPage.ToolbarItems>
@@ -12,7 +12,7 @@
   <StackLayout>
     <ListView ItemsSource="{Binding Expenses}">
       <ListView.Behaviors>
-        <blueMonkey:ItemSelectedToCommandBehavior Command="{Binding UpdateExpenseCommand}"/>
+        <behaviors:ItemSelectedToCommandBehavior Command="{Binding UpdateExpenseCommand}"/>
       </ListView.Behaviors>
       <ListView.ItemTemplate>
         <DataTemplate>
