+++ client/BlueMonkey/BlueMonkey.Views/ReportListPage.xaml
@@ -2,7 +2,7 @@
 <ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:prism="clr-namespace:Prism.Mvvm;assembly=Prism.Forms"
-             xmlns:blueMonkey="clr-namespace:BlueMonkey;assembly=BlueMonkey"
+             xmlns:behaviors="clr-namespace:BlueMonkey.Views.Behaviors;assembly=BlueMonkey.Views"
              prism:ViewModelLocator.AutowireViewModel="True"
              x:Class="BlueMonkey.Views.ReportListPage"
              Title="Report List">
@@ -11,7 +11,7 @@
   </ContentPage.ToolbarItems>
   <ListView ItemsSource="{Binding Reports}">
     <ListView.Behaviors>
-      <blueMonkey:ItemSelectedToCommandBehavior Command="{Binding UpdateReportCommand}"/>
+      <behaviors:ItemSelectedToCommandBehavior Command="{Binding UpdateReportCommand}"/>
     </ListView.Behaviors>
     <ListView.ItemTemplate>
       <DataTemplate>
