+++ client/BlueMonkey/BlueMonkey.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlueMonkey", "BlueMonkey\BlueMonkey.csproj", "{3FD586F1-4A0F-448D-918B-5F83E1B2449C}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlueMonkey.Droid", "BlueMonkey.Droid\BlueMonkey.Droid.csproj", "{4A20AC8F-DFB8-43E5-9DF9-EDC8B22A1A22}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlueMonkey.iOS", "BlueMonkey.iOS\BlueMonkey.iOS.csproj", "{E5BE7934-7D40-461B-992C-87FCB617CC64}"
@@ -63,6 +61,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlueMonkey.Transaction.Unit
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlueMonkey.Views", "BlueMonkey.Views\BlueMonkey.Views.csproj", "{8B1B81AD-B948-4EDB-94F5-E6C60DF89E51}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BlueMonkey.Application", "BlueMonkey.Application\BlueMonkey.Application.csproj", "{B5669777-4D9C-434A-97D8-7B156D87DB41}"
+EndProject
 Global
 	GlobalSection(SharedMSBuildProjectFiles) = preSolution
 		BlueMonkey.ExpenseServices.Azure.Shared\BlueMonkey.ExpenseServices.Azure.Shared.projitems*{4a20ac8f-dfb8-43e5-9df9-edc8b22a1a22}*SharedItemsImports = 4
@@ -84,30 +84,6 @@ Global
 		Release|iPhoneSimulator = Release|iPhoneSimulator
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Ad-Hoc|Any CPU.ActiveCfg = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Ad-Hoc|Any CPU.Build.0 = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Ad-Hoc|iPhone.ActiveCfg = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Ad-Hoc|iPhone.Build.0 = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Ad-Hoc|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Ad-Hoc|iPhoneSimulator.Build.0 = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.AppStore|Any CPU.ActiveCfg = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.AppStore|Any CPU.Build.0 = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.AppStore|iPhone.ActiveCfg = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.AppStore|iPhone.Build.0 = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.AppStore|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.AppStore|iPhoneSimulator.Build.0 = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Debug|iPhone.ActiveCfg = Debug|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Debug|iPhone.Build.0 = Debug|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Release|iPhone.ActiveCfg = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Release|iPhone.Build.0 = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
-		{3FD586F1-4A0F-448D-918B-5F83E1B2449C}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
 		{4A20AC8F-DFB8-43E5-9DF9-EDC8B22A1A22}.Ad-Hoc|Any CPU.ActiveCfg = Release|Any CPU
 		{4A20AC8F-DFB8-43E5-9DF9-EDC8B22A1A22}.Ad-Hoc|Any CPU.Build.0 = Release|Any CPU
 		{4A20AC8F-DFB8-43E5-9DF9-EDC8B22A1A22}.Ad-Hoc|Any CPU.Deploy.0 = Release|Any CPU
@@ -644,6 +620,30 @@ Global
 		{8B1B81AD-B948-4EDB-94F5-E6C60DF89E51}.Release|iPhone.Build.0 = Release|Any CPU
 		{8B1B81AD-B948-4EDB-94F5-E6C60DF89E51}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
 		{8B1B81AD-B948-4EDB-94F5-E6C60DF89E51}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Ad-Hoc|Any CPU.ActiveCfg = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Ad-Hoc|Any CPU.Build.0 = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Ad-Hoc|iPhone.ActiveCfg = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Ad-Hoc|iPhone.Build.0 = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Ad-Hoc|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Ad-Hoc|iPhoneSimulator.Build.0 = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.AppStore|Any CPU.ActiveCfg = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.AppStore|Any CPU.Build.0 = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.AppStore|iPhone.ActiveCfg = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.AppStore|iPhone.Build.0 = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.AppStore|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.AppStore|iPhoneSimulator.Build.0 = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Release|iPhone.Build.0 = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{B5669777-4D9C-434A-97D8-7B156D87DB41}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -670,5 +670,6 @@ Global
 		{969FB643-5584-4CDA-8238-4F3957CC123F} = {9445563F-8865-4122-8EE4-AA192F6C3AE7}
 		{7BFE197E-58E4-442F-9CBA-7E3DB0672EE2} = {9445563F-8865-4122-8EE4-AA192F6C3AE7}
 		{8B1B81AD-B948-4EDB-94F5-E6C60DF89E51} = {37AA648F-9440-4E3D-8091-302447D68EB9}
+		{B5669777-4D9C-434A-97D8-7B156D87DB41} = {6D8B156D-6918-4C98-8B45-A68F7DAEBC22}
 	EndGlobalSection
 EndGlobal
