--- client/BlueMonkey/BlueMonkey/packages.config
@@ -1,30 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="CommonServiceLocator" version="1.3" targetFramework="portable45-net45+win8" />
-  <package id="Microsoft.Azure.Mobile.Client" version="3.0.3" targetFramework="portable45-net45+win8" />
-  <package id="Microsoft.Bcl" version="1.1.10" targetFramework="portable45-net45+win8" />
-  <package id="Microsoft.Bcl.Build" version="1.0.21" targetFramework="portable45-net45+win8" />
-  <package id="Microsoft.Net.Http" version="2.2.29" targetFramework="portable45-net45+win8" />
-  <package id="Newtonsoft.Json" version="6.0.4" targetFramework="portable45-net45+win8" />
-  <package id="Prism.Core" version="6.3.0-pre1" targetFramework="portable45-net45+win8" />
-  <package id="Prism.Forms" version="6.3.0-pre1" targetFramework="portable45-net45+win8" />
-  <package id="Prism.Unity.Forms" version="6.3.0-pre1" targetFramework="portable45-net45+win8" />
-  <package id="ReactiveProperty" version="3.2.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Collections" version="4.0.11" targetFramework="portable45-net45+win8" />
-  <package id="System.ComponentModel.Annotations" version="4.1.0" targetFramework="portable45-net45+win8" />
-  <package id="System.ObjectModel" version="4.0.12" targetFramework="portable45-net45+win8" />
-  <package id="System.Reactive" version="3.0.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Reactive.Core" version="3.0.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Reactive.Interfaces" version="3.0.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Reactive.Linq" version="3.0.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Reactive.PlatformServices" version="3.0.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Reflection" version="4.1.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Reflection.Extensions" version="4.0.1" targetFramework="portable45-net45+win8" />
-  <package id="System.Runtime" version="4.1.0" targetFramework="portable45-net45+win8" />
-  <package id="System.Runtime.Serialization.Primitives" version="4.1.1" targetFramework="portable45-net45+win8" />
-  <package id="System.Runtime.Serialization.Xml" version="4.1.1" targetFramework="portable45-net45+win8" />
-  <package id="System.Xml.ReaderWriter" version="4.0.11" targetFramework="portable45-net45+win8" />
-  <package id="Unity" version="4.0.1" targetFramework="portable45-net45+win8" />
-  <package id="Xam.Plugin.Media" version="2.6.1" targetFramework="portable45-net45+win8" />
-  <package id="Xamarin.Forms" version="2.3.3.175" targetFramework="portable45-net45+win8" />
-</packages>
